src/pyproject.toml
[tool.poetry]
name = "csci338app"
version = "0.0.0"
description = "App for CSCI338"
authors = ["Semmy Purewal <semmypurewal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.104.1"}
uvicorn = "^0.23.2"
aiohttp = "^3.8.5"
pytest = "^7.4.3"
aiofiles = "^23.2.1"
httpx = "^0.25.1"
asyncpg = "^0.29.0"
sqlalchemy =  "^2.0.23"
psycopg2 = "^2.9.9"
faker = "^20.0.3"
python-dotenv = "^1.0.0"
pytest-asyncio = "^0.21.1"
trio = "^0.23.1"
sqlalchemy-utils = "^0.41.1"
pytest-trio = "^0.8.0"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "^5.12.0"
flake8 = "^6.1.0"

[tool.isort]
profile = "black"
known_first_party = ["fastapi","uvicorn"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
